最終確認

+++ practice/src/main/java/com/example/demo/repository/OrderHistoryRepository.java
@@ -9,11 +9,12 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entity.OrderHistory;
 
 @Repository
-public interface OrderHistoryRepository extends JpaRepository<OrderHistory, Long> {
+public interface OrderHistoryRepository extends JpaRepository<OrderHistory, Integer> {
 
-    // メソッド名を修正して、関連するエンティティのフィールド名を使用する
-	Optional<OrderHistory> findByProduct_IdAndStore_Id(Integer productId, Integer storeId);
+    // 商品IDと店舗IDでOrderHistoryを検索するメソッド
+    Optional<OrderHistory> findByProduct_IdAndStore_Id(Integer productId, Integer storeId);
 
-    // こちらは適切に定義されています
-	 List<OrderHistory> findByStore_Id(Integer storeId);
+    // 店舗IDでOrderHistoryのリストを取得するメソッド
+    List<OrderHistory> findByStore_Id(Integer storeId);
+
 }
